producer/requirements.txt
confluent-kafka[avro]==2.3.0
PyYAML==6.0.1
